Use CSL to format citations
Pandoc supports using the Citation Style Language to format citations,
which removes any need to biber or biblatex for the citations.

I have created a citation style which I think is particularly useful for
presentations, including the author, title and journal within the
citation. This style is based on the American Chemical society theme,
and the modern-language-association-6th-edition-note csl which supports
the footnotes.