refactor: Switch to using structopt and paw for cli
This uses the structopt and paw tools for command line argument parsing
which hugely simplify the code which needs to be written to define and
parse the arguments.

This is primarily an exercise in trying out structopt and paw, with the
conclusion that it is really simple, although there are complications
with enums, though not that much more than already exists within clap.